+++ react/components/InfoCard/index.js
@@ -97,6 +97,8 @@ const InfoCard = ({
   htmlId,
   textMode,
   linkTarget,
+  fetchpriority,
+  preload,
 }) => {
   const {
     hints: { mobile },
@@ -242,6 +244,12 @@ const InfoCard = ({
                 style={{ objectFit: 'cover' }}
                 alt={formatIOMessage({ id: callToActionText, intl })}
                 data-testid="half-image"
+                fetchPriority={fetchpriority}
+                {...(preload
+                  ? {
+                      'data-vtex-preload': 'true',
+                    }
+                  : {})}
               />
             </LinkWrapper>
           </div>
@@ -290,6 +298,8 @@ MemoizedInfoCard.defaultProps = {
   textMode: textModeTypes.TEXT_MODE_HTML.value,
   linkTarget: '_self',
   callToActionLinkTarget: '_self',
+  fetchpriority: 'auto',
+  preload: false,
 }
 
 MemoizedInfoCard.schema = {